Add abstract base template class, refactor ShineLikeDiamond to extend base template class
